+++ order-service/src/main/resources/application.properties
@@ -21,8 +21,6 @@ stripe.webhook.secret=${STRIPE_WEBHOOK_SECRET}
 # Micro Service
 product.service.name=product-service
 product.service.path=/api/v1/product
-product.inventory.service.name=product-inventory-service
-product.inventory.service.path=/api/v1/product/inventory
 
 # RabbitMQ
 spring.rabbitmq.port=5672
