+++ product-service/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.application.name=payment-service
+spring.application.name=product-service
 
 # Profiles
 spring.profiles.active=${SPRING_PROFILE:local}
